refactor: move page header component
