+++ app/src/features/layout/index.ts
@@ -1,1 +1,2 @@
 export * from './DefaultLayout'
+export * from './PageHeader'
