+++ app/src/pages/IssuePage.tsx
@@ -4,7 +4,7 @@ import { EditOutlined, StopOutlined } from '@ant-design/icons'
 import { ISSUE_STATUS_UI } from '../components/Status'
 import { renderDate } from '../data/util'
 import { IssueIcon } from '../components/ItemIcon'
-import { PageHeader } from '../components/PageHeader'
+import { PageHeader } from '../features/layout'
 import { Error404 } from '../features/ui'
 import { useParams } from 'react-router-dom'
 import { TicketLink } from '../components/TicketLink'
