+++ app/src/pages/decision-details.tsx
@@ -4,7 +4,7 @@ import { EditOutlined } from '@ant-design/icons'
 import { DECISION_STATUS_UI } from '../components/Status'
 import { renderDate } from '../data/util'
 import { DecisionIcon } from '../components/ItemIcon'
-import { PageHeader } from '../components/PageHeader'
+import { PageHeader } from '../features/layout'
 import { Error404 } from '../features/ui'
 import { useParams } from 'react-router-dom'
 import { TicketLink } from '../components/TicketLink'
